split-module api
