--- .eslintrc
@@ -1,10 +1,0 @@
-{
-  "extends": "opentable",
-  "globals": {
-    "describe": true,
-    "test": true,
-    "it": true,
-    "jest": true,
-    "expect": true
-  }
-}
