+++ packages/oc-template-handlebars/constants/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  messages: {
+    legacyComponent:
+      "The component can't be rendered because it was published with an older OC version"
+  }
+};
