+++ packages/oc-template-handlebars/getCompiledTemplate.js
