+++ packages/oc-template-handlebars/getInfo.js
@@ -1,0 +1,8 @@
+const packageJson = require('./package.json');
+
+module.exports = () => ({
+  type: packageJson.name,
+  version: packageJson.version,
+  dependencies: packageJson.dependencies,
+  externals: [packageJson.externals['handlebars']]
+});
