+++ packages/oc-template-handlebars/index.js
@@ -1,10 +1,8 @@
-const compile = require('./compile');
 const getCompiledTemplate = require('./getCompiledTemplate');
 const render = require('./render');
 const getInfo = require('./getInfo');
 
 module.exports = {
-  compile,
   getCompiledTemplate,
   render,
   getInfo
