+++ packages/oc-template-handlebars/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "oc-template-handlebars",
+  "version": "5.0.0",
+  "description": "OC-Template-Handlebars",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/opencomponents/oc-template-handlebars.git"
+  },
+  "keywords": [],
+  "author": "Nick Balestra <nick@balestra.ch>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/opencomponents/oc-template-handlebars/issues"
+  },
+  "dependencies": {
+    "handlebars": "4.0.6"
+  },
+  "homepage": "https://github.com/opencomponents/oc-template-handlebars#readme",
+  "externals": {
+    "handlebars": {
+      "global": "Handlebars",
+      "url": "https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.6/handlebars.runtime.min.js"
+    }
+  }
+}
