--- src/constants/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  messages: {
-    legacyComponent: 'The component can\'t be rendered because it was published with an older OC version'
-  }
-};
